+++ .editorconfig
@@ -8,13 +8,5 @@ indent_style = space
 indent_size = 2
 end_of_line = lf
 
-[*.{js,json,vue}]
-indent_style = space
-indent_size = 2
-
-[*.sh]
-indent_style = space
-indent_size = 2
-
 [*.md]
 trim_trailing_whitespace = false
