+++ .eslintrc.js
@@ -1,0 +1,27 @@
+module.exports = {
+  root: true,
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint', 'jest'],
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended', 'plugin:jest/recommended', 'prettier'],
+  env: {
+    browser: true,
+    es6: true,
+    node: true,
+  },
+  settings: {},
+  globals: {
+    Atomics: 'readonly',
+    SharedArrayBuffer: 'readonly',
+  },
+  parserOptions: {},
+  rules: {
+    '@typescript-eslint/explicit-function-return-type': 'off',
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
+    '@typescript-eslint/no-explicit-any': 'off',
+    '@typescript-eslint/no-var-requires': 0,
+    '@typescript-eslint/ban-ts-comment': 'off',
+    'no-prototype-builtins': 0,
+    '@typescript-eslint/no-misused-new': 0,
+    'no-undef': 0,
+  },
+}
