--- .eslintrc.json
@@ -1,31 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "es6": true,
-    "node": true,
-    "jest/globals": true
-  },
-  "plugins": [
-    "jest",
-    "prettier"
-  ],
-  "extends": [
-    "airbnb-base",
-    "plugin:vue/recommended",
-    "plugin:prettier/recommended"
-  ],
-  "globals": {
-    "Atomics": "readonly",
-    "SharedArrayBuffer": "readonly"
-  },
-  "parserOptions": {
-    "parser": "babel-eslint",
-    "ecmaVersion": 2018,
-    "sourceType": "module"
-  },
-  "rules": {
-    "import/prefer-default-export": "off",
-    "vue/html-self-closing": "off",
-    "vue/max-attributes-per-line": "off"
-  }
-}
