+++ .gitignore
@@ -70,3 +70,10 @@ TODO\.md
 /certs/cert.pem
 /certs/server.cert
 /certs/server.key
+
+/.idea
+/node_modules/
+/output/
+/dist/
+
+/yarn.lock
