--- .prettierrc.json
@@ -1,4 +1,0 @@
-{
-  "printWidth": 110,
-  "trailingComma": "all"
-}
