+++ package.json
@@ -21,59 +21,55 @@
     "url": "https://github.com/cosmos/ledger-cosmos-js/issues"
   },
   "homepage": "https://github.com/cosmos/ledger-cosmos-js",
+  "resolutions": {
+    "@ledgerhq/hw-transport": "6.27.1",
+    "@ledgerhq/devices": "6.27.1",
+    "@ledgerhq/hw-transport-node-hid": "6.27.1",
+    "@ledgerhq/hw-transport-node-hid-noevents": "6.27.1"
+  },
   "dependencies": {
-    "@babel/runtime": "^7.11.2",
-    "@ledgerhq/hw-transport": "^6.2.0",
+    "@babel/runtime": "^7.14.8",
+    "@ledgerhq/hw-transport": "6.27.1",
     "bech32": "^1.1.4",
     "ripemd160": "^2.0.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.11.6",
-    "@babel/core": "^7.11.6",
-    "@babel/node": "^7.10.5",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/preset-env": "^7.11.5",
-    "@ledgerhq/hw-transport-node-hid": "^5.25.0",
-    "@ledgerhq/hw-transport-u2f": "^5.25.3",
-    "@ledgerhq/hw-transport-webusb": "^5.25.0",
-    "@ledgerhq/logs": "^5.25.0",
-    "@vue/cli-plugin-babel": "^4.5.6",
-    "@vue/cli-plugin-eslint": "^4.5.6",
-    "@vue/cli-service": "^4.5.6",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.3.0",
+    "@babel/cli": "^7.18.10",
+    "@babel/core": "^7.18.10",
+    "@babel/node": "^7.18.10",
+    "@babel/plugin-transform-runtime": "^7.18.9",
+    "@babel/preset-env": "^7.18.9",
+    "@ledgerhq/hw-transport-node-hid": "6.27.1",
+    "@ledgerhq/hw-transport-webusb": "6.27.1",
+    "@swc/core": "^1.2.237",
+    "@types/jest": "^26.0.24",
+    "@types/ledgerhq__hw-transport": "4.21.4",
+    "@typescript-eslint/eslint-plugin": "^5.33.1",
+    "@typescript-eslint/parser": "^5.33.1",
     "bip32": "2.0.6",
     "bip39": "3.0.2",
-    "core-js": "^3.6.5",
-    "crypto-js": "4.0.0",
-    "eslint": "^7.10.0",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-config-prettier": "^6.12.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.0.2",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-vue": "^7.0.0",
-    "index.js": "^0.0.3",
-    "jest": "^26.4.2",
-    "jest-serial-runner": "^1.1.0",
-    "prettier": "^2.1.2",
-    "secp256k1": "^4.0.2",
-    "vue": "^2.6.12",
-    "vue-template-compiler": "^2.6.12",
-    "ts-jest": "^26",
-    "typescript": "^4.3.5",
     "copyfiles": "^2.4.1",
-    "ts-node": "^10.8.1"
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^28.1.3",
+    "core-js": "^3.15.2",
+    "crypto-js": "4.1.1",
+    "eslint": "^8.22.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.8.3",
+    "eslint-plugin-prettier": "^3.4.0",
+    "jest": "^26",
+    "jest-serial-runner": "^1.2.0",
+    "prettier": "^2.3.2",
+    "secp256k1": "^4.0.2",
+    "ts-jest": "26",
+    "ts-node": "^10.8.1",
+    "typescript": "^4.7.4"
   },
   "moduleDirectories": [
     "node_modules",
     "dist"
   ],
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
   "browserslist": [
     "> 1%",
     "last 2 versions"
@@ -93,9 +89,8 @@
   "scripts": {
     "build": "tsc && yarn copy-files",
     "copy-files": "copyfiles -u 0 src/**/*.proto dist/",
-
-    "test:unit": "jest -c jest.config.unit.js",
-    "test:integration": "jest -c jest.config.integration.js --runInBand --detectOpenHandles",
+    "test:integration": "yarn build && jest -t 'Integration'",
+    "test:unit": "yarn build && jest tests/basic_unit.test.ts",
     "serve": "openssl req -nodes -new -x509 -config certs/openssl.cnf -keyout certs/server.key -out certs/server.cert; openssl x509 -in certs/server.cert -out certs/cert.pem -outform PEM; vue-cli-service serve vue_example/main.js"
   }
 }
